合并

+++ zan-extension/src/core/heap.c
@@ -17,7 +17,6 @@
  */
 
 
-#include <inttypes.h>
 #include "swBaseData.h"
 #include "swMemory/memoryPool.h"
 
@@ -221,6 +220,6 @@ void swHeap_print(swHeap *heap)
     int i;
     for(i = 1; i < heap->num; i++)
     {
-        printf("#%d\tpriority=%"PRIu64", data=%p\n", i, heap->nodes[i]->priority, heap->nodes[i]->data);
+        printf("#%d\tpriority=%llu, data=%p\n", i, (long long unsigned)heap->nodes[i]->priority, heap->nodes[i]->data);
     }
 }
